moduleDir renamed to projectDir; build aborts on issue with config

+++ ProjectTemplate/Build.ps1
@@ -2,14 +2,17 @@
 .SYNOSIS
 Build a static HTML site from Markdown files.
 #>
-[string]$SCRIPT:moduleDir = Split-Path -Parent $MyInvocation.MyCommand.Path
+[string]$SCRIPT:projectDir = Split-Path -Parent $MyInvocation.MyCommand.Path
 Import-Module -Name MarkDownToHTML
 
 # JSON configuration
-$SCRIPT:config = Get-Content (Join-Path $moduleDir 'Build.json') | ConvertFrom-Json
+$SCRIPT:config = Get-Content (Join-Path $projectDir 'Build.json') | ConvertFrom-Json
+if (!$config) {
+    throw 'No build configuration found!'
+}
 
 # Location of the static HTML site to build
-$SCRIPT:staticSite = Join-Path $moduleDir $config.site_dir
+$SCRIPT:staticSite = Join-Path $projectDir $config.site_dir
 
 # Clean up the static HTML sote before build
 Remove-Item $staticSite -Recurse -Force -ErrorAction:SilentlyContinue
@@ -24,9 +27,9 @@ $SCRIPT:contentMap = @{
 	}
 }
 
-Find-MarkdownFiles (Join-Path $moduleDir $config.markdown_dir) -Exclude $config.Exclude `
+Find-MarkdownFiles (Join-Path $projectDir $config.markdown_dir) -Exclude $config.Exclude `
 | Convert-MarkdownToHTMLFragment -IncludeExtension $config.markdown_extensions `
-| Publish-StaticHTMLSite -Template (Join-Path $moduleDir $config.HTML_Template) `
+| Publish-StaticHTMLSite -Template (Join-Path $projectDir $config.HTML_Template) `
                          -ContentMap  $contentMap `
 	                     -SiteDirectory $staticSite
